Move GetSystemDefaultInterface() to Common/Network.h
